+++ tests/test_utils.py
@@ -5,7 +5,7 @@ from __future__ import absolute_import, division, print_function
 
 import pytest
 
-from packaging.utils import canonicalize_name
+from packaging.utils import canonicalize_name, canonicalize_version
 
 
 @pytest.mark.parametrize(
@@ -25,3 +25,22 @@ from packaging.utils import canonicalize_name
 )
 def test_canonicalize_name(name, expected):
     assert canonicalize_name(name) == expected
+
+
+@pytest.mark.parametrize(
+    ("version", "expected"),
+    [
+        ('1.4.0', '1.4'),
+        ('1.40.0', '1.40'),
+        ('1.4.0.0.00.000.0000', '1.4'),
+        ('1.0', '1'),
+        ('1.0+abc', '1+abc'),
+        ('1.0.dev0', '1.dev0'),
+        ('1.0.post0', '1.post0'),
+        ('1.0a0', '1a0'),
+        ('1.0rc0', '1rc0'),
+        ('100!0.0', '100!0'),
+    ]
+)
+def test_canonicalize_version(version, expected):
+    assert canonicalize_version(version) == expected